+++ dst/screens/ServiceScreen.js
@@ -11,7 +11,7 @@ import Spinner from "ink-spinner";
 import { execa } from "execa";
 import { DateTime } from "luxon";
 import { useScreen, SelectIndicator, SelectItem, runCommand } from "./Screen.js";
-import OutputBox from "./OutputBox.js";
+import OutputBox from "../components/OutputBox.js";
 const actions = [
     { label: "Start service", value: "start" },
     { label: "Stop service", value: "stop" }