+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## 0.0.4 (2026-05-28)
 
 -   IMPROVEMENT: add right-aligned line numbers to `OutputBox` output lines
+-   REFACTOR:    move `OutputBox` component from `screens/` into `components/` folder
 
 ## 0.0.3 (2026-05-28)